+++ core/member/src/main/kotlin/kr/spot/core/member/presentation/command/MemberCommandController.kt
@@ -1,56 +1,29 @@
-package kr.spot.core.member.presentation
+package kr.spot.core.member.presentation.command
 
 import io.swagger.v3.oas.annotations.Operation
-import io.swagger.v3.oas.annotations.Parameter
 import io.swagger.v3.oas.annotations.tags.Tag
 import kr.spot.common.api.ApiResponse
 import kr.spot.common.api.status.SuccessStatus
 import kr.spot.core.member.application.MemberCommandService
-import kr.spot.core.member.application.MemberQueryService
-import kr.spot.core.member.presentation.dto.request.CreateTestMemberRequest
-import kr.spot.core.member.presentation.dto.request.RegisterPreferredCategoryRequest
-import kr.spot.core.member.presentation.dto.request.UpdateMemberInfoRequest
-import kr.spot.core.member.presentation.dto.response.CreateTestMemberResponse
-import kr.spot.core.member.presentation.dto.response.GetMemberInfoResponse
-import kr.spot.core.member.presentation.dto.response.GetMemberNameResponse
-import org.springframework.web.bind.annotation.*
+import kr.spot.core.member.presentation.command.dto.request.CreateTestMemberRequest
+import kr.spot.core.member.presentation.command.dto.request.RegisterPreferredCategoryRequest
+import kr.spot.core.member.presentation.command.dto.request.UpdateMemberInfoRequest
+import kr.spot.core.member.presentation.query.dto.response.CreateTestMemberResponse
+import org.springframework.web.bind.annotation.DeleteMapping
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.PutMapping
+import org.springframework.web.bind.annotation.RequestAttribute
+import org.springframework.web.bind.annotation.RequestBody
+import org.springframework.web.bind.annotation.RequestHeader
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
 
 @Tag(name = "회원")
 @RestController
 @RequestMapping("/api/members")
-class MemberController(
+class MemberCommandController (
     private val memberCommandService: MemberCommandService,
-    private val memberQueryService: MemberQueryService
-) {
-    // ==================== Query ====================
-
-    @Operation(summary = "회원 이름 조회")
-    @GetMapping("/name")
-    fun getMemberName(
-        @Parameter(hidden = true) @RequestHeader memberId: Long
-    ): ApiResponse<GetMemberNameResponse> {
-        val name = memberQueryService.getMemberName(memberId)
-        return ApiResponse.ok(
-            GetMemberNameResponse
-                .from(name)
-        )
-    }
-
-    @Operation(summary = "회원 정보 조회")
-    @GetMapping("/info")
-    fun getMemberInfo(
-        @Parameter(hidden = true) @RequestHeader memberId: Long
-    ): ApiResponse<GetMemberInfoResponse> {
-        val member = memberQueryService.getMember(memberId)
-        return ApiResponse.ok(
-            GetMemberInfoResponse.from(
-                member
-            )
-        )
-    }
-
-    // ==================== Command ====================
-
+){
     @Operation(summary = "회원 선호 카테고리 설정")
     @PostMapping("/prefer-categories")
     fun registerPreferCategories(
